requirements.txt
Flask==2.3.3
groq==0.4.2
python-dotenv==1.0.0